requirements.txt
pandas
numpy
torch
torchvision
torchaudio
transformers
pillow
matplotlib
scipy
tqdm
pyarrow
albumentations